warehouse_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "warehouse-order-fulfillment"
version = "1.0.0"
description = "Warehouse Order Fulfillment RL Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.3",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "fastmcp>=0.1.0",
    "groq>=0.4.0",
]

[project.scripts]
server = "warehouse_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["warehouse_env", "warehouse_env.envs", "warehouse_env.server"]
package-dir = { "warehouse_env" = "." }
